refactor: move project related utils from `core.utils2` to `project` app - `core/utils2/project.py` to `project/utils/project_utils.py` - `core/utils2/projects.py` to `project/utils/projects_utils.py` - `core/utils2/project_seed.py` to `project/utils/projectseed_utils.py`
