+++ docker-app/qfieldcloud/core/tests/test_project.py
@@ -22,12 +22,12 @@ from qfieldcloud.core.models import (
     TeamMember,
 )
 from qfieldcloud.core.tests.utils import wait_for_project_ok_status
-from qfieldcloud.core.utils2 import project_seed
 from qfieldcloud.project.models import (
     SHARED_DATASETS_PROJECT_NAME,
     Project,
     ProjectSeed,
 )
+from qfieldcloud.project.utils import projectseed_utils
 from qfieldcloud.subscription.models import Subscription
 
 from .utils import set_subscription, setup_subscription_plans, testdata_path
@@ -63,7 +63,7 @@ class QfcTestCase(APITransactionTestCase):
 
         ProjectSeed.objects.create(
             project=project,
-            extent=Polygon.from_bbox(project_seed.DEFAULT_PROJECT_EXTENT),
+            extent=Polygon.from_bbox(projectseed_utils.DEFAULT_PROJECT_EXTENT),
             settings={
                 "schemaId": ProjectSeed.SETTINGS_SCHEMA_ID,
                 "basemaps": [],
