+++ docker-app/qfieldcloud/project/utils/project_utils.py
