+++ docker-app/qfieldcloud/project/utils/projects_utils.py
@@ -1,6 +1,5 @@
 from django.db.models import Q
 from django.utils.translation import gettext as _
-
 from qfieldcloud.core import invitations_utils as invitation
 from qfieldcloud.core import permissions_utils as perms
 from qfieldcloud.core.models import (
